Issue #681 - Put apt commands in non-interactive mode
When apt commands are in non-interactive mode, they accept the
default position instead of querying the user for a response. In
the scripts, if this occurs, it causes a failure. This happens in
the rare case when an apt update causes a package to be removed
when installing a new package.