Issue #623 - Add a homelabos distro/release variable
By having a homelabos distribution and release, non-standard users, like those
that use debian based distro's that are not ubuntu, can specify in their config file
for homelabos to use the custom values instead of the hardcoded defaults.

The defaults are set to the ansible values, so that no interaction is required by
the user to get a working system on ubuntu server.

Since PureOS uses debian buster repos instead of ubuntu's, this fixes the docker-ce install
issue associated with #623 once the user specifies the proper variable values.

Update Ansible version to 2.9.7 in order to support ubuntu20.04.
See:
https://github.com/ansible/ansible/pull/69161/files
https://stackoverflow.com/questions/61460151/ansible-not-reporting-distribution-info-on-ubuntu-20-04
https://www.digitalocean.com/community/tutorials/how-to-install-and-configure-ansible-on-ubuntu-20-04