Issue #680: Fix inline jinja templates to the proper quotes
If an inline jinja conditional is not properly quoted, then it inserts an extra space
at the end of the condition. This normally isn't a problem unless the resulting variable is
used inside something that interprets that space incorrectly, e.g. they don't trim whitespace
from input.

This can be fixed by wrapping the inline conditionals in quotes as suggested here:
https://stackoverflow.com/questions/58125886/inline-jinja2-if-condition-in-ansible-playbook

Then there is the problem of nested quotes, so move them to single quotes, to fix that problem
and make things more consistent.